refactor: rename files
Rename files to better be in line with other file names
